+++ index.html
@@ -211,8 +211,5 @@
       </div>
     </template>
     <script type="module" src="./pages/index.js"></script>
-    <script type="module" src="./pages/validation.js"></script>
-    <script type="module" src="./pages/ava.js"></script>
-    <script type="module" src="./pages/modal.js"></script>
   </body>
 </html>
